requirements.txt
Flask
Flask-PyMongo
Flask-Session
Authlib
requests
python-dotenv
google-generativeai
dnspython